requirements.txt
Flask==2.1.2
scikit-learn==1.1.1
pandas==1.4.2
psycopg2-binary==2.9.3
python-dotenv==1.0.1
